Ass support with pysubs2
