+++ PySubtitle/Helpers/Time.py
@@ -106,3 +106,51 @@ def TimedeltaToSrtTimestamp(time: datetime.timedelta|str|None) -> str|None:
 
     return f"{hours:02}:{minutes:02}:{seconds:02},{milliseconds:03}"
 
+def AssTimeToTimedelta(ass_time: str) -> datetime.timedelta:
+    """
+    Convert ASS time format (H:MM:SS.CC) to timedelta.
+    
+    Args:
+        ass_time: Time string in ASS format (e.g., "1:23:45.67")
+        
+    Returns:
+        datetime.timedelta: Converted time
+    """
+    try:
+        # ASS time format: H:MM:SS.CC where CC is centiseconds
+        time_match = regex.match(r'(\d+):(\d{2}):(\d{2})\.(\d{2})', ass_time)
+        if not time_match:
+            return datetime.timedelta(seconds=0)
+        
+        hours = int(time_match.group(1))
+        minutes = int(time_match.group(2))
+        seconds = int(time_match.group(3))
+        centiseconds = int(time_match.group(4))
+        
+        return datetime.timedelta(
+            hours=hours,
+            minutes=minutes,
+            seconds=seconds,
+            milliseconds=centiseconds * 10  # Convert centiseconds to milliseconds
+        )
+    except (ValueError, AttributeError):
+        return datetime.timedelta(seconds=0)
+
+def TimedeltaToAssTime(td: datetime.timedelta) -> str:
+    """
+    Convert timedelta to ASS time format (H:MM:SS.CC).
+    
+    Args:
+        td: Time delta to convert
+        
+    Returns:
+        str: Time in ASS format (e.g., "1:23:45.67")
+    """
+    total_seconds = int(td.total_seconds())
+    hours = total_seconds // 3600
+    minutes = (total_seconds % 3600) // 60
+    seconds = total_seconds % 60
+    centiseconds = int((td.microseconds / 1000) / 10)  # Convert microseconds to centiseconds
+    
+    return f"{hours}:{minutes:02d}:{seconds:02d}.{centiseconds:02d}"
+
